retry delete
