+++ assets/src/components/stacks/Stacks.tsx
@@ -41,8 +41,9 @@ import { LinkTabWrap } from '../utils/Tabs'
 import { LoadingIndicatorWrap } from '../utils/LoadingIndicator'
 
 import CreateStack from './create/CreateStack'
-import DeleteStack from './delete/DeleteStack'
 import StackEntry from './StacksEntry'
+import StackDetach from './StackDetach'
+import StackDelete from './StackDelete'
 
 export type StackOutletContextT = {
   stack: StackFragment
@@ -213,10 +214,6 @@ export default function Stacks() {
                 {stack.repository?.url}
               </div>
             </div>
-            <DeleteStack
-              stack={stack}
-              refetch={refetch}
-            />
             <KickButton
               pulledAt={stack.repository?.pulledAt}
               kickMutationHook={useKickStackMutation}
@@ -224,6 +221,14 @@ export default function Stacks() {
               tooltipMessage="Use this to sync this stack now instead of at the next poll interval"
               variables={{ id: stack.id }}
               width="max-content"
+            />
+            <StackDetach
+              stack={stack}
+              refetch={refetch}
+            />
+            <StackDelete
+              stack={stack}
+              refetch={refetch}
             />
           </div>
           <TabList
